# openapi.yaml
openapi: "3.0.3"
info:
  title: Secure Multi-Tier Platform API
  description: |
    Enterprise-grade SaaS API backend demonstrating multi-tier AWS architecture.
    Provides product catalogue management, order processing, user authentication
    via AWS Cognito (OAuth2 with PKCE), and platform health monitoring.
  version: "1.0.0"
  contact:
    name: Platform Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development (Docker Compose)
  - url: https://api.example.com
    description: Demo deployment (API Gateway + CloudFront)
...
security:
  - BearerAuth: []
paths:
...
  /products:
    get:
      operationId: listProducts
      summary: List products with pagination and optional category filter
      description: |
        Returns a paginated list of active products. Supports filtering by
        category and sorting. Results are cached using the cache-aside pattern
        with a 60-second TTL.
      tags:
        - products
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: category
          in: query
          description: Filter products by category
          schema:
            type: string
        - name: sort_by
          in: query
          description: Field to sort by
          schema:
            type: string
            default: created_at
      responses:
        "200":
          description: Paginated list of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createProduct
      summary: Create a new product
      description: |
        Creates a new product in the catalogue. Requires admin or manager role.
        Invalidates the product list cache on success.
      tags:
        - products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
            example:
              name: "Pale Ale Malt"
              description: "Premium base malt for pale ales"
              price_pence: 2500
              stock_quantity: 100
              category: "malts"
      responses:
        "201":
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions (requires admin or manager role)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
  /products/{product_id}:
    get:
      operationId: getProduct
      summary: Get a single product by ID
      description: |
        Returns a single product by its UUID. Cached with a 300-second TTL.
      tags:
        - products
      parameters:
        - name: product_id
          in: path
          required: true
          description: Product UUID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateProduct
      summary: Update a product
      description: |
        Updates an existing product. Only provided fields are changed.
        Requires admin or manager role. Invalidates relevant caches on success.
      tags:
        - products
      parameters:
        - name: product_id
          in: path
          required: true
          description: Product UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
            example:
              price_pence: 2750
              stock_quantity: 85
      responses:
        "200":
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions (requires admin or manager role)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"

    delete:
      operationId: deleteProduct
      summary: Delete (soft-delete) a product
      description: |
        Soft-deletes a product by marking it inactive. Requires admin role only.
        Invalidates relevant caches on success.
      tags:
        - products
      parameters:
        - name: product_id
          in: path
          required: true
          description: Product UUID
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Product deleted successfully (no content)
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions (requires admin role)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from AWS Cognito via OAuth2 Authorization Code
        flow with PKCE. Include the token in the Authorization header as:
        `Authorization: Bearer <token>`
  schemas:
...
    ProductCreate:
      type: object
      required:
        - name
        - price_pence
        - stock_quantity
        - category
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Product name
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Product description
        price_pence:
          type: integer
          minimum: 1
          maximum: 10000000
          description: Price in pence (max £100,000)
        stock_quantity:
          type: integer
          minimum: 0
          description: Available stock quantity
        category:
          type: string
          minLength: 1
          maxLength: 100
          description: Product category
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          nullable: true
          description: Product name
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Product description
        price_pence:
          type: integer
          minimum: 1
          maximum: 10000000
          nullable: true
          description: Price in pence
        stock_quantity:
          type: integer
          minimum: 0
          nullable: true
          description: Available stock quantity
        category:
          type: string
          minLength: 1
          maxLength: 100
          nullable: true
          description: Product category
        is_active:
          type: boolean
          nullable: true
          description: Whether the product is active
    ProductResponse:
      type: object
      required:
        - id
        - name
        - price_pence
        - stock_quantity
        - category
        - is_active
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Product UUID
        name:
          type: string
          description: Product name
        description:
          type: string
          nullable: true
          description: Product description
        price_pence:
          type: integer
          description: Price in pence
        stock_quantity:
          type: integer
          description: Available stock quantity
        category:
          type: string
          description: Product category
        is_active:
          type: boolean
          description: Whether the product is active
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
    ProductListResponse:
      type: object
      required:
        - items
        - total
        - page
        - page_size
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProductResponse"
          description: List of products for the current page
        total:
          type: integer
          description: Total number of matching products
        page:
          type: integer
          description: Current page number
        page_size:
          type: integer
          description: Number of items per page

    # --- Orders ---
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "INVALID_CREDENTIALS"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid email or password"
            request_id:
              type: string
              nullable: true
              description: Request correlation ID for tracing
            details:
              type: object
              nullable: true
              additionalProperties: true
              description: Additional error context
    ValidationErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
                description: Field location path
              msg:
                type: string
                description: Validation error message
              type:
                type: string
                description: Error type identifier
          description: List of field-level validation errors